drivers: uart: uart_xlnx_ps: update interrupt acknowledgement handling
Change the initialization when CONFIG_UART_INTERRUPT_DRIVEN is enabled to *not* clear pending interrupts.

Add work-around to initialization to send a NUL character to trigger a TX interrupt if a TX interrupt is not already pending.

Move the acknowledgement of TX interrupts out of xxx_irq_tx_ready() and into xxx_fifo_fill() to ensure that the TX IRQ is not lost in the case where the client chooses to disable TX interrupts when the client does not have any additional data to transmit.